+++ src/gfx/pscanvas.h
@@ -48,7 +48,7 @@ public:
   virtual ~PSCanvas() throw();
 
   virtual geom::vec3<double> screenFromWorld3(const geom::vec3<double>& world_pos) const;
-  virtual geom::vec3<double> worldFromScreen3(const geom::vec3<int>& screen_pos) const;
+  virtual geom::vec3<double> worldFromScreen3(const geom::vec3<double>& screen_pos) const;
 
   virtual geom::rect<int> getScreenViewport() const;
